Separated recipe library into multiple files
Also:
- Added documentation
- Moved writeToFile to recipe:
While implementing analysis.py I realized that writeToFile, Run and read
the results should be implemented in analysisRecipe level and just be
called from analysis.py library with an optional list of HBObjects.

Another thought is that HBObjects can be also implemented in recipe
level. I'm still not sure how good of an idea is this. We can implement
it later if needed.

@sariths this made  some delay to finish the whole workflow but I
thought this is a major change in the workflow and I should implement it
for 3Phase method. You will have a working version of the grasshopper
file at some point before tomorrow morning. I will send an email with
more details. The file will be under test folder >
https://github.com/ladybug-analysis-tools/honeybee-grasshopper/tree/master/test