Added Rtrace and run to grasshopper definition
I haven't implemented read results yet and run analysis is not fully
implemented but this commit provides the full workflow of running a
grid-based analysis using Honeybee API.

@sariths I made changes to Grasshopper file which I hope solves your
issue. There is a component on the top left of the script that you can
provide the address to where you have the honeybeegrasshopper library
downloaded. Let me know how it goes.