Added runTest.py for testing Grasshopper scripts
Check this issue for more info:
https://github.com/ladybug-analysis-tools/honeybee-core/issues/7#issuecomment-194472827

@sariths try running runTest.py. It will open Rhino, execute all the
grasshopper files in the folder, closes Rhino and compile the .log
files.

There is a component in .gh file that generates the log file once it's
opened. You need to copy-paste the component in your code and provide
expected values from the script. I'm casting everything to string to
make it easier to test. using this workflow we can create test files and
run the tests all together.