First try for merging geometry operations
This is the first prototype to address
https://github.com/ladybug-analysis-tools/honeybee-core/issues/8
To address this issue I added a config.py which indicates the execution
platform and imports related geometry libraries.

You can see the use case in geometryoperation.py where each method
checks for the platform first and based on the platform uses the related
library.
Next step is to test this workflow inside dynamo and in case it works
fine in Dynamo we should rename this repository and use it both for
Grasshopper and Dynamo.

I'm thinking of qbee, queenbee, xbee, bee or honeybeeX as the new name.