Used SkyMtx with an hour for skyVector
During the workshop it was clear that running perl scripts on Windows
machines is not the best idea. Regradless of the issue with finding the
path to OpenStudio installation many of the users couldn't run the
skyVector successfully. To address this issue skyVector component is
also using skyMtx but generates the results for a single hour of the
year.

This should resolve
https://github.com/ladybug-analysis-tools/honeybee/issues/117 and
https://github.com/ladybug-analysis-tools/honeybee/issues/114